refactor: rename `modules/pipeline.py` to `modules/animation_pipeline.py`
